move root.go to subdir cmd/cli
